mutation_partition: Add support for specifying continuity
This will allow expressing lack of information about certain range of
rows (including the static row). This will be used in cache to
determine if information in cache is complete or not.

For each entry, rows_entry::continuous() determines whether the range
between that entry and the previous one (both ends exclusive) is
continuous or not. The range after the last rows_entry is assumed to
be continuous.  It can be marked as discontinuous by inserting a so
called dummy entry (rows_entry::dummy()) with appropriate key. Keys
corresponding to regular rows_entries are assumed to be
continuous. Keys corresponding to dummy entries are assumed to be
discontinuous.

Continuity information is ignored by instance equality. It's also
transient, not preserved by serialization.

Addition (apply) of two fully-continuous instances gives a
fully-continuous instance.  When adding two not fully-continuous
instances, the information other than continuity is added the same way
as for fully-continuous instances. Resulting continuity is currently
unspecified though.

Addition of continuity could be well-defined as set intersection, with
any key in the result being continuous if and only if it is continuous
in both instances.  That implies that a regular row which falls into
discontinuous range of another instance would be dropped. The reason
why it is not defined as such is that it is some effort to implement
and this functionality is not used yet on mutation_partition level.
The semantics mentioned here are implemented on partition_entry level,
by partition_entry::apply_to_incomplete().

Based on the following commits from seastar-dev.git haaawk/partial_cache/v5:

  4a5bf75 - Piotr Jastrzebski : mutation_partition: introduce dummy rows_entry
  773070e - Piotr Jastrzebski : mutation_partition: add continuity flag to rows_entry