mvcc: Make each version have independent continuity
This change is a preparation for introducing row-level eviction, such that entries
can be evicted from older versions without having to touch other versions.

Currently continuity flags on entries are interpreted relative to the
combined view merged from all entries. For example:

 v2:                  <key=2, cont=1>
 v1: <key=1, cont=1>

In v2, the flag on entry key=2 marks the range (1, 2) as
continuous. This is problematic because if the old version is evicted, continuity
will change in an incorrect way:

   v2:                  <key=2, cont=1>

Here, the range (-inf, 1) would be marked as continuous, which is not true.

To solve this problem, we change the rules for continuity
interpretation in MVCC. Each version will have its own continuity,
fully specified in that version, independent of continuity of other
versions. Continuity of the snapshot will be a union of continuous
ranges in each version.

It is assumed that continuous intervals in different versions are non-
overlapping, except for points corresponding to complete rows, in
which case a later version may overlap with an older version
(overwrite). We make use of this assumption to make calculation of the
union of intervals on merging easier. I make use of the above
assumption in mutation_partition::apply_monotonically().

MVCC population of incomplete entries already almost maintains the
non-overlapping invariant, because population intervals correspond to
intervals which are incomplete in the old snapshot. The only change
needed is to ensure that both population bounds will have dummy
entries in the latest version. Population from memtables doesn't mark
any intervals as continuous, so also conforms. The only change needed
there is to not inherit continuity flags from the old snapshot,
effectively making the new version internally discontinuous except for
row points.

The example from the beginning will then become:

 v2: <key=1, cont=1, dummy>  <key=2, cont=1>
 v1: <key=1, cont=1>

When marking a range as continuous with some rows present only in older
versions, dummy entries will be inserted in the latest version, which marks ranges
corresponding to those rows as discontinuous. When merged with older versions,
the complete row jumps into the gap represented by a dummy entry and it would
form a fully continuous range. It would be also correct to copy the rows from
older versions into the latest version, but less efficient.

This change doesn't yet insert dummy entries for population bounds,
because cache is not preprated to encounter dummy entries at
clustering row positions, it inserts empty rows instead. This is
correct as there is no partial eviction yet. This will be changed in a
later patch.