tests: cache: Improve debuggability of failures of test_concurrent_reads_and_eviction
Write failure messages comparing actual version with each of the
expected version separately, so that the tests/mutation_diff script
can be run on each of them to see the diffs between the actual
mutation and each of the expected versions.

Also, write diffs between each of the expected versions, to see
incremental changes made to the cache.

It's almost impossible to debug assertion failures without this.