gdb: Switch task_histogram to use the span_checker
It can be that page::pool is != nullptr and page::offset_in_span is 0
for a page which is inside a large allocation span (live or
dead). This may lead to misqualification of that span as belonging to
a small allocation pool and interpreting its contents as if it
contained small objects.

Only the first page of a span contains reliable information. This
patch changes the code to use the span_checker, which knows the real
boundaries of spans and exposes reliable information via the span
object.

Another problem was that the command scanned dead spans as well.
This is no longer the case after this patch.

I've seen this command report thousands of no longer live sstable
writers and various continuations because of those problems.

Fixes #4367