sstables: ka/la: reader: Make sure push_ready_fragments() does not return yes when partition_end is pending but not emitted
Currently, if there is a fragment in _ready on _out_of_range was set
(because consume_row_end was called), push_ready_fragments() would
return proceed::yes. This will cause the reader to invoke the parser,
which will stop after encountering the next partition. This is
problematic once we make consume_row_start() emit partiton_start
directly, because we will want to assume that all fragments for the
previous partition are emitted by then. If they're not, then we'd emit
partition_start without emitting partition_end for the previous
partition. The fix is to make sure that push_ready_fragments() returns
yes only when there's nothing left to push.